# src/openapi/system-spec.yaml
openapi: 3.0.3
info:
  title: HAXcms System API
  description: Control-plane API for authentication, lifecycle/admin, discovery/meta, and system settings/configuration.
  version: 0.0.0
servers:
  - url: /
    description: System API base path
...
paths:
...
  /system/api/v1/session/user:
    get:
      tags:
        - session
      operationId: sessionUserGet
      summary: Return active user profile and dashboard preference data
      security:
        - bearerAuth: []
          userTokenHeader: []
      responses:
        "200":
          description: Active user profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - session
      operationId: sessionUserPost
      summary: Return active user profile and dashboard preference data
      security:
        - bearerAuth: []
          userTokenHeader: []
      responses:
        "200":
          description: Active user profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /system/api/v1/sites:
    get:
      tags:
        - lifecycle
      operationId: listSites
      summary: List all sites available to the authenticated user
      security:
        - bearerAuth: []
          userTokenQuery: []
      responses:
        "200":
          description: Site listing payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - lifecycle
      operationId: createSite
      summary: Create a new site
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteActionRequest"
      responses:
        "200":
          description: Site creation response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /system/api/v1/sites/{siteName}/clone:
    parameters:
      - $ref: "#/components/parameters/SiteName"
    post:
      tags:
        - lifecycle
      operationId: cloneSite
      summary: Clone an existing site
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteActionRequest"
      responses:
        "200":
          description: Site clone response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/sites/{siteName}/archive:
    parameters:
      - $ref: "#/components/parameters/SiteName"
    post:
      tags:
        - lifecycle
      operationId: archiveSite
      summary: Archive a site
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteActionRequest"
      responses:
        "200":
          description: Site archive response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/sites/{siteName}/download:
    parameters:
      - $ref: "#/components/parameters/SiteName"
    post:
      tags:
        - lifecycle
      operationId: downloadSite
      summary: Generate and return a downloadable site archive reference
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteActionRequest"
      responses:
        "200":
          description: Site download response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/sites/{siteName}/download-skeleton:
    parameters:
      - $ref: "#/components/parameters/SiteName"
    post:
      tags:
        - lifecycle
      operationId: downloadSiteSkeleton
      summary: Generate and return a reusable skeleton for a site
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteActionRequest"
      responses:
        "200":
          description: Site skeleton response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/sites/{siteName}/save-as-template:
    parameters:
      - $ref: "#/components/parameters/SiteName"
    post:
      tags:
        - lifecycle
      operationId: saveSiteAsTemplate
      summary: Save a site as a reusable system template
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteActionRequest"
      responses:
        "200":
          description: Save-as-template response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/system/app-store:
    get:
      tags:
        - settings
      operationId: generateAppStore
      summary: Generate app-store provider manifest for the editor
      security:
        - appStoreTokenQuery: []
          siteTokenQuery: []
      responses:
        "200":
          description: App-store payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/system/status:
    get:
      tags:
        - settings
      operationId: systemStatusGet
      summary: Return system status and environment diagnostics
      security:
        - bearerAuth: []
          userTokenQuery: []
      responses:
        "200":
          description: System status data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - settings
      operationId: systemStatusPost
      summary: Return system status and environment diagnostics
      security:
        - bearerAuth: []
          userTokenQuery: []
      responses:
        "200":
          description: System status data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/settings/api-keys:
    get:
      tags:
        - settings
      operationId: getApiKeys
      summary: Return configured API keys and provider statuses
      security:
        - bearerAuth: []
          userTokenQuery: []
      responses:
        "200":
          description: API key settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - settings
      operationId: saveApiKeysPost
      summary: Update API key settings
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeysSettings"
      responses:
        "200":
          description: API key update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      tags:
        - settings
      operationId: saveApiKeysPatch
      summary: Update API key settings
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeysSettings"
      responses:
        "200":
          description: API key update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/settings/media:
    get:
      tags:
        - settings
      operationId: getMediaSettings
      summary: Return media and upload configuration
      security:
        - bearerAuth: []
          userTokenQuery: []
      responses:
        "200":
          description: Media settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - settings
      operationId: saveMediaSettingsPost
      summary: Update media and upload configuration
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaSettings"
      responses:
        "200":
          description: Media settings update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      tags:
        - settings
      operationId: saveMediaSettingsPatch
      summary: Update media and upload configuration
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaSettings"
      responses:
        "200":
          description: Media settings update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/settings/enabled-skeletons:
    post:
      tags:
        - settings
      operationId: saveEnabledSkeletonsPost
      summary: Update enabled skeleton configuration
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnabledCollectionSettings"
      responses:
        "200":
          description: Enabled skeletons update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      tags:
        - settings
      operationId: saveEnabledSkeletonsPatch
      summary: Update enabled skeleton configuration
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnabledCollectionSettings"
      responses:
        "200":
          description: Enabled skeletons update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/settings/schema-files/operations:
    post:
      tags:
        - settings
      operationId: schemaFileOperation
      summary: Perform schema file operation in system configuration storage
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Schema file operation response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/settings/enabled-themes:
    post:
      tags:
        - settings
      operationId: saveEnabledThemesPost
      summary: Update enabled theme configuration
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnabledCollectionSettings"
      responses:
        "200":
          description: Enabled themes update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      tags:
        - settings
      operationId: saveEnabledThemesPatch
      summary: Update enabled theme configuration
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnabledCollectionSettings"
      responses:
        "200":
          description: Enabled themes update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /system/api/v1/system/blocks:
    get:
      tags:
        - settings
      operationId: systemBlocksGet
      summary: Return available system blocks
      security:
        - bearerAuth: []
          userTokenQuery: []
      responses:
        "200":
          description: Available block list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - settings
      operationId: systemBlocksPost
      summary: Return available system blocks
      security:
        - bearerAuth: []
          userTokenQuery: []
      responses:
        "200":
          description: Available block list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
    SiteName:
      in: path
      name: siteName
      required: true
      schema:
        type: string
      description: Machine name of the site
...
  responses:
    BadRequest:
      description: Request payload was invalid or incomplete
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Forbidden:
      description: Authenticated user is not authorized for this route
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    userTokenQuery:
      type: apiKey
      in: query
      name: user_token
    userTokenHeader:
      type: apiKey
      in: header
      name: X-HAXCMS-User-Token
    siteTokenQuery:
      type: apiKey
      in: query
      name: site_token
    appStoreTokenQuery:
      type: apiKey
      in: query
      name: appstore_token
  schemas:
    ApiEnvelope:
      type: object
      properties:
        status:
          type: integer
        data: {}
      additionalProperties: true
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            message:
              type: string
...
    SiteActionRequest:
      type: object
      properties:
        site:
          type: object
      additionalProperties: true
    ApiKeysSettings:
      type: object
      description: API key settings payload
      additionalProperties: true
    MediaSettings:
      type: object
      description: Media settings payload
      additionalProperties: true
    EnabledCollectionSettings:
      type: object
      description: Enabled item map payload
      additionalProperties: true
...